requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
PyMySQL==1.1.0
cryptography==42.0.0
Pillow==10.2.0
requests==2.31.0
huggingface_hub==0.30.2
